renamed backstop files to 'test' cleaned up config files & created main config, that is spread into / overwritten by local or live config files moved aem username / password to .env file
