+++ .gitignore
@@ -94,9 +94,8 @@ site-template
 theme/.env
 
 ### BackstopJS Files ###
-theme/backstop_data/bitmaps_test/
-theme/backstop_data/html_report/
-theme/backstop_data/backstop-login-info.js
+theme/test_data/bitmaps_test/
+theme/test_data/html_report/
 
 ### ESLint ###
 .eslintcache
