+++ theme/package.json
@@ -17,11 +17,11 @@
     "watch": "run-p watch:*",
     "watch:color-scheme": "parcel watch src/color-scheme.ts",
     "watch:theme": "parcel watch src/theme.ts",
-    "test-live-config": "backstop test --config=live.backstop.config.js",
-    "test:local": "backstop test --config=local.backstop.config.js",
+    "test-live-config": "backstop test --config=live.test.config.js",
+    "test:local": "backstop test --config=local.test.config.js",
     "test:live": "run-p live test-live-config",
-    "test:update-local": "backstop approve --config=local.backstop.config.js",
-    "test:update-live": "backstop approve --config=live.backstop.config.js"
+    "test:update-local": "backstop approve --config=local.test.config.js",
+    "test:update-live": "backstop approve --config=live.test.config.js"
   },
   "devDependencies": {
     "@adobe/aem-site-theme-builder": "5.2.1",
