+++ theme/test_data/bitmaps_reference/backstop_default_Home_0_document_0_phone.png
