+++ theme/test_data/bitmaps_reference/backstop_default_mobile-menu_0_document_0_phone.png
