+++ theme/test_data/engine_scripts/playwright/clickAndHoverHelper.js
