+++ theme/test_data/engine_scripts/playwright/interceptImages.js
