+++ theme/test_data/engine_scripts/puppet/clickAndHoverHelper.js
