+++ theme/test_data/engine_scripts/puppet/interceptImages.js
