+++ theme/test_data/engine_scripts/puppet/loadCookies.js
