+++ theme/test_data/engine_scripts/puppet/onReady.js
